fix(client): correct AgentParams bond type to uint96 + Transfer event fallback
The ServiceManagerToken implementation uses IService.AgentParams with
uint96 bond (not uint256). This caused a function selector mismatch
(0x89fffda9 vs 0xe42cdd7c) making every create() call revert.

Also adds ERC721 Transfer event fallback for parsing service ID from
the creation tx, since the CreateService event topic may not match.

Service creation now works on Anvil fork (service 537 created).
Bootstrap proceeds to service_activated step.

+++ client/src/earning/contracts.ts
@@ -151,7 +151,7 @@ export const SERVICE_MANAGER_ABI = [
       { name: 'agentIds', type: 'uint32[]' },
       { name: 'agentParams', type: 'tuple[]', components: [
         { name: 'slots', type: 'uint32' },
-        { name: 'bond', type: 'uint256' },
+        { name: 'bond', type: 'uint96' },
       ]},
       { name: 'threshold', type: 'uint32' },
     ],
